Swift: move `TargetTrapDomain` to `TargetDomains` in `infra`
