+++ swift/extractor/infra/BUILD.bazel
@@ -6,6 +6,7 @@ swift_cc_library(
     hdrs = glob(["*.h"]),
     visibility = ["//swift:__subpackages__"],
     deps = [
+        "//swift/extractor/config",
         "//swift/extractor/infra/file",
         "//swift/extractor/trap",
         "//swift/third_party/swift-llvm-support",
