+++ swift/extractor/infra/TargetDomains.h
